+++ src/model/game/animals/guardanimals/Cat.java
@@ -1,7 +1,7 @@
 package model.game.animals.guardanimals;
 
 public class Cat extends GuardAnimal{
-    public Cat(float x, float y) {
+    public Cat(int x, int y) {
         super(x, y);
     }
