+++ src/model/game/animals/guardanimals/Dog.java
@@ -1,7 +1,7 @@
 package model.game.animals.guardanimals;
 
 public class Dog extends GuardAnimal{
-    public Dog(float x, float y) {
+    public Dog(int x, int y) {
         super(x, y);
     }
 }
