+++ src/model/game/animals/guardanimals/GuardAnimal.java
@@ -5,7 +5,7 @@ import model.game.animals.Animal;
 public abstract class GuardAnimal extends Animal {
     public static final int CAT_PRICE = 400;
     public static final int DOG_PRICE = 200;
-    public GuardAnimal(float x, float y) {
+    public GuardAnimal(int x, int y) {
         super(x, y, 1);
     }
 }
