+++ src/model/game/animals/predatoranimals/Lion.java
@@ -1,7 +1,7 @@
 package model.game.animals.predatoranimals;
 import static changes.PredatorAnimal.LION;
 public class Lion extends PredatorAnimal{
-    public Lion(float x, float y) {
+    public Lion(int x, int y) {
         super(x, y, LION);
     }
 }
