+++ src/model/game/animals/predatoranimals/Tiger.java
@@ -3,7 +3,7 @@ package model.game.animals.predatoranimals;
 import static changes.PredatorAnimal.TIGER;
 
 public class Tiger extends PredatorAnimal{
-    public Tiger(float x, float y) {
+    public Tiger(int x, int y) {
         super(x, y, TIGER);
     }
 }
